apps/api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "jobpilot-api"
version = "1.0.0"
description = "JobPilot AI Backend API - 24/7 Job Agent Platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "asyncpg>=0.29.0",
    "redis>=5.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[argon2]>=1.7.4",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "slowapi>=0.1.9",
    "email-validator>=2.1.0",
    "pypdf2>=3.0.0",
    "python-docx>=1.1.0",
    "python-magic-bin>=0.4.14; sys_platform == 'win32'",
    "python-magic>=0.4.27; sys_platform != 'win32'",
    "aiofiles>=23.2.0",
    "feedparser>=6.0.10",
    "apscheduler>=3.10.4",
    "aiosqlite>=0.19.0",
    "scikit-learn>=1.3.0",
    "reportlab>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "hypothesis>=6.92.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]
